refactor: extract shared content to separate file
Avoids confusing circular import
