+++ websites/M/MyFigureCollection.net/pages/ClassifiedPage.ts
@@ -1,5 +1,5 @@
 import { getButton } from '../util.js'
-import { BasePage, strings } from './index.js'
+import { BasePage, strings } from './base.js'
 
 export class ClassifiedPage extends BasePage {
   override async executeView(presenceData: PresenceData, _id: string): Promise<boolean> {
