+++ websites/M/MyFigureCollection.net/pages/ShopPage.ts
@@ -1,5 +1,5 @@
 import { getButton, getCurrentLink, getThumbnail, getTitle } from '../util.js'
-import { BasePage, strings } from './index.js'
+import { BasePage, strings } from './base.js'
 
 export class ShopPage extends BasePage {
   override async executeView(presenceData: PresenceData, _id: string): Promise<boolean> {
