+++ websites/M/MyFigureCollection.net/pages/TagPage.ts
@@ -1,5 +1,5 @@
 import { getTitle } from '../util.js'
-import { BasePage, strings } from './index.js'
+import { BasePage, strings } from './base.js'
 
 export class TagPage extends BasePage {
   override async executeView(presenceData: PresenceData, _id: string): Promise<boolean> {
